validate: Split application unit tests to separate file
Move application-specific validation unit tests (TestValidatedDRPC*,
TestValidatedVRGState*, TestValidatedProtectedPVC*) from command_test.go
to application_validated_test.go. These tests validate individual
functions without running the full command flow.

This keeps application_test.go available for the full command flow tests
when we split those next, and the file will move to
pkg/validate/application/ when we split the commands.
